services/backend/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
bidict==0.22.0
click==8.1.2
cycler==0.11.0
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.1
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
fonttools==4.31.2
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.16.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==0.21.0
alembic==1.12.1
redis==5.0.1
Flask-Caching==2.1.0
prometheus_client